# Cache negative dentries for faster negative lookups